src/mtgcards/api/serializers.py
from mtgcards.api.models import Card
from rest_framework import serializers


class CardSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Card
        fields = '__all__'
