src/requirements.txt
gunicorn
django
djangorestframework
requests
psycopg2-binary
pillow 
opencv-python
ijson
tqdm